Began re-re-re-<times X>-doing the GUI visualization framework.
This time, however, there's a new twist. Instead of completely separating out
non-abstracted visualizations from abstracted (statistical) visualizations,
instead everything is grouped together into one Visualization class.

The Visualization class is currently operating by simply being nothing other
than a control interface. Of course, this is changable by the type of
visualization, but in essence, all it does is display a QImage that is
generated by a VisualizationThread.

The VisualizationThread is combined with VisualizationData instances to render
the said QImage. In essence, VisualizationThread calls something along the
lines of VisualizationData::render_self(this), to render itself upon the
given QImage using an interface provided by the VisualizationThread. The
interface will be made up of methods such as draw_rectangle(), which could be
used to create a bar graph.

However cut and clear it may seem at the moment through the eyes of fatigue,
that is bound to change in the harsh light of day, with the clarity that only
morning can bring. The "facepalms", to use a colloquial term..