Implemented Timestamp::ms_until() . . .
And I finally figured out why SessionOverview::started() was getting called;
there is a naming conflict here. DataThread is supposed to emit a started()
signal, which is caught by the Session, and then passed onto the
SessionOverview. However, DataThread inherits QThread, which has a signal
called started() . . .

So, a renaming is required here, methinks . . .