Re-implemented TimeSlider.
For some reason, my brain did not realize that TimeSlider was actually a
TimeSpinBox . . .

Now, I need to rename some signals. In DataThread(), the signal start() should
be renamed to . . . hmm. data_start(), perhaps? I suppose I'll just prepend on
data_ to the beginning of the signal names, at least as far as the DataThread
signals go . . .