Almost implemented the ActiveBlocks visualization . . .
Except for two things: first off, the QPainter that the ActiveBlocksData::
render_to() method creates claims that the QImage is invalid. Second of all,
the data that ActiveBlocksRequest is gathering seems a bit skewed sometimes.
For example, it claims that the amount of active blocks is much lower than it
actually is, and also claims that the number of leaked blocks is falling as
the program executes . . . both of which are major issues . . .

I'm not sure what's wrong with those, but I should try to figure out exactly
what is wrong . . . why the ActiveBlocksRequest isn't working is much more
serious, methinks. I'll take another look.