The only possibility here: the tree is not set up correctly . . .
If I show exactly what is happening in count_blocks(), it shows that it is
only acting upon the state of the tree in snapshot #1. Also, the nice qDebug()
statement is showing that the new blocks are always being added to the first
snapshot's head element.

In other words, I need to pin down exactly where the tree is going wrong . . .