A problem with the tricky part . . .
In this case, the "tricky" part is the event-based accuracy version of
SnapshotList::get_snapshot_for(). This is the primary problem: the temporary
snapshots created (in this case, created with negative IDs, starting from -1)
need to be treated normally -- events need to be able to be applied to them,
etc. However, the original tree must be left untouched, the parents
unmodified. That is where the tricky part comes in . . .

I suppose I could clone the entire tree, but that would be naught but a waste
of memory . . . looks like I need to do some trickery with the tree. But what
kind of trickery? . . .