A solution: get rid of parent nodes.
This means that two things must be done when inserting a new node into the
tree: first of all, the head node must be correct (e.g. for the snapshot that
the tree is currently being modified by), and the last "correct" node that
found must be stored. Why? Currently the parent pointer in the tree nodes is
used for two things: to streamline the process of "marking" the tree, and to
find the newly-created "head" node in AllocEvent when the tree is marked.

Therefore, it should be possible to completely remove the parent node, and
thus solve the problem introduced with the temporary snapshots.