SnapshotList::get_snapshot_for() now properly applies all events correctly.
This means that it is now possible to get millisecond-accurate representations
of the program's memory, as the ActiveBlocksVisualization now properly
demonstrates.

There is one issue, though: the temporary snapshots are never freed or cached.
This means that a huge amount of memory is being utilized to find such data.

Also, I have noticed a bug in the VisualizationRequest system . . . for some
reason or another, the Visualization requests never actually get through, or
at least the DataThread never sends any responses or something similar. I'll
have to look into this some more.