That seems to have fixed it . . .
The QTimer apparently didn't like being started and stopped so much. Instead,
the request_queue_timer in DataThread is now a straight 500 ms.

However, this brings me to another problem: after ~5k temporary snapshots, the
GUI was using ~2GB of memory . . . I think I need to free them . . .

Freeing the temporary snapshots is quite easy, actually: all that has to be
done is to go through the list and find the nodes marked changed with the ID
of the current temporary snapshot, push them onto a list/queue, then free
them . . .