The current rendering system will not entirely work in the future . . .
The reason being: currently, each VisualizationData is completely unaware of
any other VisualizationDatas. Picture, for example, a line graph: each data
element has to know about its direct neighbours (actually, for a line graph,
only one of the neighbours needs to be known, but there are other
visualizations where both need to be known). For a line graph, the problem is
easily solved: add a "draw_point" class into the associated VThread . . .

But, what about the time/address space graph I mentioned earlier? In such a
visualization, there are large gaps in the address space (in glibc, when
allocating a block larger than MMAP_THRESHOLD bytes, the memory is actually a
private mmap, which has an address close to 0x8000000000000000 . . . a typical
address from a non-mmap address is 0x600000). In other words, such gaps need
to be found ahead of time, so that the data can be displayed in the correct
place.

Also, another thing I need to implement is a grid or something similar on the
graphs; graphs are rather useless without a reference point . . .