One possible method to attempt to implement in the rendering system . . .
In essence, the system is nothing more than another inheritance tree. Rather
than painting a line on a QImage from (0, 0) to (0, 200) to indicate a value
of 200, instead the VisualizationData creates a DataPoint of (0, 200), which
the VisualizationThread then creates a dynamically-scaling graph out of,
automatically including the "squiggly lines" that denote a period of the axis
that is empty. Then, once the address space/time graph is implemented, it
simply says, a DataRectangle from (20.0, 0x601010) to (30.0, 0x601018), and
the VThread does its job in the same way as before. These DataPoints etc. will
be stored inside the VisualizationRequest, of course, since the DataPoints are
rather specfic . . . though DataPoints sound remarkably similar to VData
instances . . .

So, I'll simply revamp the current VisualizationData tree a little . . .