Added --version flag to the monitor, removed the missing config files warning.
Also continued working on the VisualizationRender* classes. It's a tad
convoluted, but, what of my work is not? Convolution should be my middle name,
at least judging by Aesalon . . .

In essence, this is how the rendering system is going to work. First, we start
off with an already-generated VisualizationData (I'm not explaining how those
are accquired. Go back a few commits if you're really interested . . .) . . .

The VData creates a VisualizationRenderObject with the VRenderRange that is
associated with it (that bit is new, by the way). Actually, to be more
specific, a subclass of VisualizationRenderObject is created. The VRObject is
then given to a VRenderer instance, where it is stored inside a QList (or
something similar). The VRenderer class then, when an rerender is requested,
calculates the VRRanges (i.e. calculates the regions of the visualization with
no data, and adds a "squiggly line" to denote that fact), and gives the
VRObjects coordinates to render themselves to.