Note: this version segfaults upon execution.
For some reason, the GUI is segfaulting on line 52 of VisualizationThread.cpp,
now that the VRenderer class is actually being used. I'm not sure what the
issue is, but I can definitely say that I am too tired to find it tonight.

I'll try to solve it in the morning. I highly suspect that it is a problem
with the multiple threads . . . but that wouldn't make sense; it is an easily
reproducable problem, not one that only happens occasionally . . .