Changed the VisualizationCanvas to use a QScrollArea.
Now it is possible for the visualization to exceed the size of the canvas; and
indeed, a pair of scrollbars appears when this is the case.

However, I think I should switch from using QImages to QPixmaps; QPixmaps are
easier to paint onto the screen, and I'm not loading/saving the images very
often.