Fixed a few small bugs that were making large amounts of trouble in the GUI.
Primarily was the one in NetworkReceiver, where, annoyingly enough, addresses
with (& 0xff > 0x80) were causing an overflow somewhere, and turning the
address into 0xffffffff . . .

Secondly, FreeEvent wasn't properly marking the tree as changed. Actually, it
was, but it was utilizing the old head node in its traversal. As one could
guess, that was not exactly healthy for the tree.