Implemented a count_events() function in SnapshotList.
However, that is not the answer to the speed problems. Rather than using
linear searches everywhere, there is a better way . . .

In essence, new-Snapshot-per-n-events needs to be implemented. Rather than
taking a snapshot every i ms, as it is currently, instead snapshots should be
taken every n events, or i ms, whichever is shorter. That reduces the chances
of a situation like I am currently in with the intentionally-inefficient
fibonacci benchmark, which has ~106k events over a time period of less than
six seconds. So, rather than having ~50k events per snapshot, instead have a
mere 1k events, short enough that a linear search will not have a huge speed
impact, unlike having 106k events . . .

To put it simply, I have a new goal for the next little while: to eliminate
all linear searches with more efficient methods. If that includes a double
binary tree in SnapshotList for accessing snapshots via their ID as well as
their associated timestamp, then so be it.

Come to think of it, a necessity introduced by taking snapshots frequently is
a higher resolution timer. On my system, Q{Date,}Time has an accuracy of 18.2
milliseconds. Or, at least, it appears to. As a result, I need to re-implement
Timestamp to have a time source of a higher accuracy . . .