Replaced the linear search in get_closest_snapshot with a binary-style search.
I had a thought while reading the Qt documentation: rather than using a raw
QPixmap, why not use something slightly higher-level? Such as, oh, how about
QPicture? From the manual pages for Qt, "The QPicture class is a paint device
that records and replays QPainter commands." In other words, a SVG-style
image, which is precisely what I want for the visualizations: an SVG image, of
which a specific part can be rendered at any given time.