Replaced another linear search in SnapshotList with a more efficient system.
I've given some thought to exactly how I would like the rendering system to be
(ideally!), and I've come to the conclusion that an SVG-style (e.g. vector
graphics) rendering system would be the best idea . . .

So, I can sort of implement this using a QPicture. QPictures have a clipping
rectangle built-in, and I can simply scale the QPicture, set the clipping
rectangle, then paint the QPicture onto the VCanvas -- no hefty Q{Image,
Pixmap}s required, which will dramatically save on memory usage and speed.

Only disadvantage I can see is the lack of double-buffering, but it is
entirely possible that QPainter has some sort of double-buffering built-in, as
long as I don't clear the screen beforehand, like I currently am.

The grid will need to be re{calculated,drawn} as well, but the grid is the
least of my worries at the moment, I think.