Changed VRenderer so it only creates one QPainter instance per update.
I think, however, that I will create a QPainterPath of the visualization,
which will be renderer directly onto a QWidget, or, rather, a customized
QAbstractScrollArea. That will solve some of the most pressing issues, and it
also means that a hefty Q[Image|Pixmap] does not need to be passed around.