Re-implemented ActiveBlocksRequest to use the new EventVisitor system.
This makes it much more efficient, but means it produces an ActiveBlocksData
instance for each event in the specified time. While nice to have such
accuracy, I think I'm going to have to limit it to ~5000 points of data per
graph . . . it takes far too long and too much memory to display the full
graph for something like fib_test. (~250MB, if you're interested. Not half bad
for 212k events (106k allocations, but reallocations currently count as two
events: a free and an allocation) . . .).

I think the grid might be the next most important thing, as I have mentioned
somewhere in the past . . . I'm not entirely certain how to go about it,
though . . .