Added warranty disclaimers to the beginning of each source file.
This is in an attempt to meet all of the requirements of the GNU GPLv3, as it
states in the section entitled, "How to Apply These Terms to Your New
Programs".

This commit brings the line count of Aesalon to over 10,000. This makes it
approximately four and a half times larger than any other project I have ever
written, and the second-largest project that I have worked on . . .