Began implementing a scope transfer system again.
This time, I will try to do it properly . . .

In essence, whenever a new scope is encountered, a ScopeEvent is sent with
the scope name -- this may include source file names as well, and that scope
is then assigned an identification number. When an allocation event occurs,
the event observer finds the return address (and indeed perhaps the scope),
then passes the ID of the scope that the return address is in, and also sends
an offset along. Currently, the offset will be unused, but eventually denote a
line number. If offset is nonzero, then the associated scope is assumed to be
a file name.

Also added support for 'make doc', which generates doxygen documentation.