A conclusion regarding the QGraphicsScene . . .
As convienient as it may be, I do not believe it is the optimal solution. I
suspect that I may be able to provide a more streamlined solution for the
generation of visualizations. If I render the visualizations myself, onto a
QImage, then I daresay it may end up being slightly faster. As a secondary
result, it will bring the processing out of the Qt rendering thread, where the
major speed delays are currently introduced, resulting in an unresponsive GUI.

As such, I think it may be worth attempting to create a custom system of
displaying the visualizations, rather than using a pre-existing solution.