I think I may have found a solution to the rendering speed.
In essence, I have done a little thinking on the problem, and come up with
what appears to be a feasible solution. The first step is to change
VisualizationDisplay into something slightly more flexible. One of the more
major issues is going to be caching the rendered images, but that's another
problem. For a more detailed breakdown, I will list what will happen upon the
user shifting the view of a visualization slightly, in this case, assume it is
shifted upwards 20 pixels. The VisualizationDisplay immediately moves the
content of the current QImage (or QPixmap, perhaps) downwards 20 pixels,
resulting in an immediate change for the user. Then, the VDisplay asks the
VRenderer to render a QImage of that 20 pixel strip, with the associated
VRPoints provided as well.

The VRenderer, running in its own thread, then finds the VData that intersect
that region of the QImage, and tells them to re-render themselves. Once the
QImage representing the updated region has been rendered, it is sent back to
the VDisplay, where it is subsequently blitted onto the VRenderer's QImage.
The temporary QImage is then destroyed (or perhaps cached if I get around to
that bit). That eliminates the VCanvas as well, I believe.

Well, it looks good in writing. How it looks in code is a completely different
question . . .