Removed VisualizationCanvas.
I think I now have a fairly decent system of rendering/updating the VDisplay.
The first step is to split the displayed QImage into three: one for the left,
one for the middle, and one for the right-hand side. When the display is being
rendered in real-time, the whole display is not updated, only the right
QImage. In other words, the display has a fixed width during real-time
updating. When the right panel is filled, then the panels are swapped.
Assuming {left,centre,right} => {0, 1, 2}, then {0,2} and {0,1} are swapped,
and the original 0, now 2, is filled with white.

As for the grid system . . . in the VisualizationDisplay, not only the above
three QImages are stored: a fourth QImage, the size of the previous three
added together horizontally is stored, which is a transparent QImage (do
QImages support transparency? I will have to look that up . . .) that is
blitted on top of the main three (or perhaps underneath, depending on where
the grid is rendered: on top of, or below the main content of the
visualization). This QImage will be updated whenever the viewport is updated.

Then, the VisualizationDisplay can set parameters such as the history length,
update interval, and other such fun stuff. Once I get this all working, it
should, in theory (note that word . . .), make Aesalon a tad more useful than
it currently is.

Only one problem I can see with this, and that is click-detection. But that's
another problem, one for another time.