Removed VisualizationRequest and all references to it.
In essence, VisualizationRequest will be replaced with a simple VRDataRange,
since that is the function it is fulfilling now . . .

Now, I need to come up with a system of easily storing the VData that are
generated in such a manner that I can request all VData for a specific
timestamp. I think a tree would work best -- it's a similar problem to the
tree used to store block data -- but I'm unsure what type of tree -- perhaps
a true quadtree, since there are two dimensions to VData . . . the problem is
the density visualization, where VData can span a huge time and/or data range.

Because of this, a standard quadtree will not quite work. A modified quadtree,
perhaps, storing a list of all items within each quadrant, (with some fancy
code to not further split a quadrant if there is no gain to doing so), could
work . . .