Mostly implemented a rendering framework independent of QGraphicsScene.
I've rethought out not using QGraphicsScene: and my thought regarding click-
detection is not exactly valid: basically, when a click is detected (or a
keypress, or whatever it is that is used to indicate that something should be
about to display the allocation scope of the block the mouse cursor is
currently over . . .), simply call VRenderer::resolve_point(), and then use
the SnapshotList tree to find which block it is referencing, by virtue of the
calculated timestamp and address from the image coordinates.

In essence, this is how the framework operates: the Controller is the place to
request a rendering, the VRenderer just does the actual rendering itself. Each
Controller instance is within its own thread, which is how the multithreading
aspect of the GUI is introduced, and the advantage over QGraphicsScene. The
renderer keeps track of the actual size of the current visualization, and can
also update said ranges without re-rendering anything (for example, this will
be used in real-time rendering when the left quarter/third/half/whatever is
chopped off, and the rest is updated as normal). The Display does nothing
other than render QImages, and the Visualization does nothing other than
contain a few widgets, the VDisplay among them.

If I have gone and hopelessly confused anyone trying to comprehend how the new
system works, then know this: you're not alone.