Began re-implementing the density visualization.
Now, instead of a series of unwieldy create_*() functions, a
VisualizationFactory is used instead, to get around the limitations of using
pure virtual functions inside C++ constructors.

Two things remain: first, a method of creating the specific VFactory. Second,
a system to actually render a region of data -- the framework is completed,
but nothing actually utilizes it yet . . .