Began implementing a system to actually use the current framework . . .
So, for a brief overview of exactly what each Visualization class does . . .
Visualization is nothing more than a GUI element, a class to draw together all
of the various classes involved with rendering a visualization.
VisualizationDisplay does nothing more than display a QImage at a given
offset and to provide a navigational interface. VisualizationRenderer is where
most of the hard work is done, it is here that data is actually rendered.
VisualizationController is simply an interface for a VisualizationUpdater to
use. VisualizationUpdaters emit signals, render_region() and move_display(),
which the VController uses to control the VRenderer. VUpdaters are created by
the VisualizationSettings dialog, which is created and launched from the
original Visualization instance. The VController is in its own thread, which
is how the VRenderer is executed outside of the main Qt/GUI thread.

Next up: to properly implement SingleUpdater, a VUpdater derivative that does
nothing more than emit a move_display() signal, then a render_region() signal.
This behaviour is identical to the setup the QGraphicsScene version used.