I have a decision to make now . . .
Are all Timestamps in the GUI stored as offsets or as absolute values? I can
either store the absolute value of each timestamp, e.g. the seconds and
nanoseconds since 23:59:59 on the 31st of December, 1969, or I can store them
as seconds and nanoseconds since the beginning of the execution of the
program.

There are advantages and disadvantages to each. The advantage of storing
timestamps as relative is obvious: the start time does not need to be
propagated throughout the program, and makes calculations on the visualization
side a tad easier. On the other hand, storing them relatively is a
disadvantage for the DataThread: the timestamps generated by the monitor are
absolute, not relative, and therefore additional work must be done per event
received.

However, events are only received once, and may be processed many times . . .