Problem with the data elements of a VDataRange . . .
The RealtimeUpdater needs to be able to change the VDataRange that the
currently-rendered QImage is representing. As such, there is the nice signal
called move_display() that does exactly that. However, the RealtimeUpdater has
no control over exactly what the user is viewing. As such, it needs to be able
to change the time range without changing the data range . . .