Redesign time.
The current design is not very good, unfortunately. It needs to be redone out
on paper/whiteboard first, this time . . .

If you think about it, none of the classes have a clearly defined role, and
they all interact in seemingly random ways . . . both are recipies for
insanity in the developer.