I believe I have come up with a design that should work . . .
There are three main classes: Viewport, Renderer, and Controller. In essence,
a Viewport displays a Canvas object, (Canvas is simply a wrapper around a
DataRange and a QImage), a Renderer is a class that creates Canvas objects,
and a Controller manages the Renderer.

In software engineering terms, there are three primary goals of this design:
to have the ability to render a visualization, to be able to display a
rendered visualization, and to be able to trans{late,form} the rendered
visualizations. The Renderer class handles the rendering aspect, the Viewport
the displaying, and the Controller handles the transformation and translation.