A thought regarding Canvas objects . . .
Perhaps I should duplicate how QGraphicsScene paints objects? In other words,
for those not familliar with QGraphicsScene operates, create objects that
render themselves onto a QPaintDevice as required? There are advantages and
disadvantages to implementing a Canvas in this manner . . .

A clear advantage: Canvases do not need to be scaled, nor do they lose any
accuracy at any time. The disadvantage is, of course, the huge amount of
objects that will be allocated . . .

But still, I like the idea. Now, to implement it . . .