Began implementing the Canvas class.
There is one change from the definition of the Canvas class in the previous
commit: Canvas objects store instances of CanvasObjects, which are much like
QGraphicsItems. This means that the Viewport needs to have a separate thread
to render onto a QImage, since I will assume that Qt::
WA_PaintOutsidePaintEvent is not available . . . ugh.

Not much I can do about it, though. There's worse fates.