And the density visualization works once more.
Well, at least it partially works . . . only when a full visualization is
requested does it actually visualize anything. But whatever, the framework is
at least operational now.

Next up: moving and scaling the canvas. That, and replacing the ugly hackish
method used to generate a Canvas with all objects within a given time frame.