Mostly re-implemented the density visualization's functionality.
Added a nice VisualizationFactory object that creates other the other
visualization-specific objects. These currently include Renderer and
AxisFormatter, a nice new class that allows the visualization to format a
DataPoint however it likes . . .

I should mention that the speed increase over the QGraphicsScene is phenomenal
-- it only takes ~250ms or less to render the fib_benchmark density
visualization, compared to approximately 5 seconds for the QGraphicsScene.

Now, for real-time updating of said visualizations . . .