Fixed one of the bugs, at least . . .
The GUI can now properly receive the data produced by the new overload
library. However, when the monitor is connected to a GUI, the speed of the
monitored program decreases dramatically -- to three seconds, from 300 ms.

The only idea I can come up with is that the monitor, GUI, and monitored
program together are taking up more CPU, which reduces the speed of all three.
Also, on a side note, during the execution of the fibonacci benchmark, an
average of 700KB/s of data is being transferred through the network device.
That is rather a lot . . .