Fixed the overload library to properly resolve calloc().
Do note that the solution is rather hack-ish: the offset of malloc() is
calculated in the monitor and put into an environment variable. The overload
library then calculates the address that libc is loaded into, then adds the
monitor-calculated offset onto that address to resolve malloc() . . .

It should work, but it might be Linux-specific. I'm not sure.