Fixed a memory leak in the Canvas-based system.
Canvases and CObjects were being allocated but never freed. In other words,
the GUI was using about 1 GB of memory after a few minutes of real-time
updating . . .

Now, I definitely need to implement a better system of real-time updating.
For example, partial paintings in the CanvasPainter, etc. Anything to speed up
the GUI a little . . .