Implemented most of a good scope transfer system.
In essence, rather than attempting to transfer scopes on-demand, adding more
processing requirements to the monitor, the GUI will have a list of all the
symbols, which it will then find the symbols for the scope addresses for.

When I say "most", I mean a "functional-but-flawed" version. In essence, the
system works perfectly for blocks less than 2**(64 - tree_depth) bytes large,
due to the method used to look up the block to find its allocation
information . . .

Which means I am back to the exact same problem that I was a little while ago,
and that is how to efficiently store ranges of 2D data . . .