Tracked down a bug relating to certain events not being received.
Most of the blocks in a large portion of the visualizations were not marked as
having a release time or scope, when the program (and valgrind) clearly stated
they were released.

Basically, what is happening is this: the pipe buffer (which is 64KB,
according to pipe(7)), is not being completely emptied when the monitored
program exits. The signal that the program has terminated is received well
before the pipe buffer is cleared, and so the events still waiting to be
processed are never received . . .

The solution to this is to move OverloadParser into its own thread, I think.
That should clear it up . . .