Mostly implemented RenderedCanvas.
For some reason, though, nothing is ever being displayed by Viewport. I do not
know why, but somewhere, something is not working correctly. I'll have to see
if I can find it, I guess . . .

On an unrelated note, another problem: for programs that have a long backlog
of events (e.g. those that utilize the pipe's buffer to its fullest extent),
the timestamp they finish at and the timestamp associated with the completion
of the program's execution varies by about three seconds . . . which is
suboptimal.