As much as I dislike it, I see only one solution.
Canvas objects need to be cloned. In essence, what this means is that the
CanvasPainter must be given a copy of the current Viewport::local_canvas.
Which means that another linear-time operation is introduced . . .

But I see no way around it, short of implementing a versioned red-black or
other self-balancing tree. Come to think of it, that would actually work. It
does not actually have to be a binary tree, as nothing other than sequential
lookups are ever attempted. So, that means that a temporary copy of a portion
of a Canvas object can be made, without creating a copy of a QList or
something similar.

You know, that might actually work.