Mostly implemented a linked-list to attempt to solve the multithreading issue.
However, there still appears to be a problem: when switched to multithreading
mode, there are still some white gaps in the output, where there should not be
any. However, that is a problem for tomorrow, not toda-err, sorry. Later on
this morning, not now. I am officially too tired to think straight.