Partial rendering appears to be complete. On both axis.
Tested on the fib_benchmark output, the difference is appreciable. It is
actually possible to maneuver around effectively, unlike how it was before.

However, there appears to be an off-by-one error that is producing slightly
strange results in several situations: probably a rounding error. At the
moment, however, I am content to let it lie.

I have more interesting things to do.