Began re-implementing the ELF parser to support 32-bit ELF files.
There are a few issues, however, that will likely result in a complete
redesign. The most major issue is the fact that ELF32 files use Elf32_Word
instances as the standard, and ELF64 files use Elf64_Word . . . not entirely
certain how I will be solving that one.

On the other hand, I have also come up with a redesign for the networking
system, in an attempt to reduce bandwidth on 32-bit systems. The new iteration
is very similar to the current, with three exceptions: first of all, the
header byte has an additional flag somewhere, denoting the word size (32 or
64). Second, all addresses and other such platform-size-specific data are sent
in the "native" size -- e.g. 32 bits for 32-bit platforms, 64 bits for 64-bit
platforms. Third, all "type" data is stored within the header byte. I do
realize this limits the total amount of event types to 128 (7 bits, since one
is taken for 32/64-bit distinction), but since there are currently only 6
used, I feel that it is not a large disadvantage.