A thought regarding the ELF parser . . .
The ELF parser is too specific, i.e. it is not abstracted enough from the data
it represents. Instead of calling it the ELF parser, it should instead be
called the resolver, after all, since that is its primary function . . . to
parse ELF data to find the section offsets and symbol {name/,offset/,size}s.

Thus, the ELF parser (and all associated uses) need to be revamped into one
interface, along with the DWARF parser and any other interfaces that are non-
dynamic (meaning that the same request will always provide the same response,
unlike the ptrace interface . . .), such as the disassembler interface . . .
though, come to think of it, that is one part of the monitor that is sadly
derelict. I think I will keep it, though . . . could be useful in the future.

This will provide a unified interface that will be magnitudes easier to port,
since the ELF parser will not have to be re-implemented for POSIX systems that
use an executable format other than ELF, and so will the debugging information
parsing be abstracted somewhat from the format used (DWARF, STABS, etc.) . . .