Wait, what was I thinking? The storage-offset system is actually faster . . .
In essence, I believe the sample I used to determine speed was not quite
unbiased. When I re-run the tests now, the speed difference is sometimes as
much as 100% -- 100 ms for the StorageManager-based system, and 200 ms for the
standard allocation system . . . but the only thing that can really be at
fault is my averaging program . . . I suppose it must not have been working
correctly.

Well, one thing is for certain: I will not be using the current StorageManager
setup. It is too unwieldy . . . instead, the Analyzer will internally use a
StorageManager-like system, but will not expose the StorageOffsets, etc. to
the rest of the monitor.

The rest of the monitor will, in all likelihood, not be able to utilize a
system like this, since the objects created there will be allocated, released,
and changed too often for it to work very well. Since I do want to get Aesalon
to a reasonable stage of completion this decade, I think I will stay away from
memory fragmentation . . .