I am fairly sure that the monitor will work on 32-bit platforms now.
Of course, it still remains to be tested. That, and the GUI needs to implement
support for the new dual-length Words . . . though I think the GUI will keep
using quint64's everywhere; it is just easier.

Do note: overload-path is now a path, not a file. overload-path should point
to a directory that contains "aesalon_overload_{32,64}.so", with 32/64-bit
versions as required.