Re-implemented the Analyzer's storage method again.
In essence, everything is now either a StorageAttribute or a simple
StorageOffset. It is a fairly simple system, and an interface is provided so
that the storage system is transparent to the rest of the monitor.

However, there is a bug with storing values, for some reason that I cannot
conceive. I store a value, it goes away when I try to retreive it. Simple
situation, but I have no clue why it disappears . . .