Removed depenances upon a hardcoded path to libc.
Instead, the --libc-path argument value is used. This makes it quite a bit
easier to find the correct libc . . . means that you do not have to recompile
Aesalon if you compiled it with the wrong libc_path.

In fact, I think I will eliminate the LIBC_PATH entry from build.config
entirely, and rely solely upon the run-time configuration to find the path to
libc . . .