As per usual, there is another major hurdle to overcome . . .
This time it is a rather . . . fundamental problem, unfortunately. In essence,
I never considered what would happen if a process forked itself. The multi-
threading aspect of that does not bother me, I would assume that all UNIX pipe
writing operations are atomic for a buffer of less than 64 bytes. (It would,
of course, be a different situation if the buffer size was 4096 bytes, but I
digress.) Therefore, the sending/receiving code would be in good stead, at
least as far as that goes. The main issue here is: how will the data be
stored? The data should not be presented in one visualization, either, because
many of the addresses will be overlapping (remember, these are forks/execs,
not separate threads) . . . that needs to be an option, I think.

Perhaps the "monitor" event setup currently used could be adapted to a
"process" event, as in process started/stopped? The pid of the created process
could be used to distinguish easily between created processes. All that would
require is a simple modification to add a word representing the process ID
number the event represents . . .

Which leads the way for another visualization. Each thread or process (I'm not
sure how I could differentiate between threads. There is something else to
research . . .) is represented by a polygon. The taller the polygon, the more
memory the process was using at that particular time; the wider it is, the
longer the process was running for. Come to think of it, this sort of
visualization would not work directly for threads, since memory can be
allocated in one thread and released in another, but something similar could
be applied, I suppose.

This sort of information would definitely help determine how a process forks
itself off . . . in fact, it would be the first step towards providing non-
specific visualizations of data relating to a program. Aesalon could -- with
enough work -- be applied as a more generic visualization tool for programs on
POSIX-based operating systems. Signals received, system calls produced, child
processes created, and other such useful information could be presented in a
visual form.

Definitely something to think about.